Fix for bug in corePKCS11 PAL layer that caused reset after OTA
A bug got introduced in the PAL of the corePKCS11 port, which caused a reset during NVS initialisation due to incorrect usage of the critical section entry / exit APIs. Zero was passed instead of a valid mux.
The use of critical section to protect the NVS initialisation has been replaced with a mutex, for the following reasons:
i) Interrupts get disabled in the critical section, and NVS initialisation is a flash operation.
ii) A flash operation may take long, and busy-waiting using a spinlock might trigger the interrupt WDT.

This commit fixes https://github.com/espressif/esp-aws-iot/issues/112